Restructured CLI
