+++ cmd/wfcli/parse.go
@@ -1,0 +1,55 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/fission/fission-workflows/pkg/api/workflow/parse/yaml"
+	"github.com/golang/protobuf/jsonpb"
+	"github.com/urfave/cli"
+)
+
+var cmdParse = cli.Command{
+	Name:        "parse",
+	Aliases:     []string{"p"},
+	Usage:       "parse <path-to-yaml> ",
+	Description: "Parse YAML definitions to the executable JSON format (deprecated)",
+	Action: func(c *cli.Context) error {
+
+		if c.NArg() == 0 {
+			panic("Need a path to a yaml workflow definition")
+		}
+
+		for _, path := range c.Args() {
+
+			fnName := strings.TrimSpace(path)
+
+			f, err := os.Open(fnName)
+			if err != nil {
+				panic(err)
+			}
+
+			wfDef, err := yaml.Parse(f)
+			if err != nil {
+				panic(err)
+			}
+
+			wfSpec, err := yaml.Transform(wfDef)
+			if err != nil {
+				panic(err)
+			}
+
+			marshal := jsonpb.Marshaler{
+				Indent: "  ",
+			}
+			jsonWf, err := marshal.MarshalToString(wfSpec)
+			if err != nil {
+				panic(err)
+			}
+
+			fmt.Println(jsonWf)
+		}
+		return nil
+	},
+}
